refactor: HttpVersion 관련 로직 수정
