+++ tomcat/src/test/java/org/apache/catalina/resource/ResourceHandlerTest.java
@@ -10,7 +10,7 @@ import org.apache.coyote.http11.data.HttpMethod;
 import org.apache.coyote.http11.data.HttpRequest;
 import org.apache.coyote.http11.data.HttpResponse;
 import org.apache.coyote.http11.data.HttpStatusCode;
-import org.apache.coyote.http11.data.HttpVersion;
+import org.apache.coyote.HttpVersion;
 import org.apache.coyote.http11.data.MediaType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
